Add ability to disable ACL for system streams
This allows our subscriptions to function properly against
not-yet-created category streams. Synce category streams are considered
"system" streams by EventStore, the ACL for reading system streams has
to be relaxed. Otherwise, 401 forbiddens are returned in place of 404
unauthorized. This is so that unauthenticated users cannot learn what
category streams do and do not exist.